leam-base.commons-lang.mid-342.idx-39910.6.mutant

+++ commons-lang/src/main/java/org/apache/commons/lang3/mutable/MutableByte.java
@@ -318,9 +318,8 @@ public class MutableByte extends Number implements Comparable<MutableByte>, Muta
      * @return the numeric value represented by this object after conversion to type double.
      */
     @Override
-    public double doubleValue() {
-        return value;
-    }
+        public double doubleValue() {
+    return (double)value;    }
 
     /**
      * Gets this mutable as an instance of Byte.
